feat(storage): create bucket state
